Adjust naming conventions, add noop facts
The naming convention standardizes all the facts to begin with
agent_specified, to help group and indicate what they are for.

As these facts are intended to help with testing workflow, it makes
sense to also include whether or not --noop was specified. This would
allow users to do things like allow nodes to request environments
besides their own to do test runs against, but only if they were
operating in no-op mode.